requirements.txt
fastapi==0.103.0 
Uvicorn==0.23.2
starlette==0.27.0
prisma==0.10.0
requests==2.31.0
PyJWT==2.8.0
coloredlogs==15.0.1